// src/app/interfaces/blog-image.ts
export interface BlogImage {
    public_id:         string;
    version:           number;
    signature:         string;
    width:             number;
    height:            number;
    format:            string;
    resource_type:     string;
    created_at:        string;
    tags:              any[];
    bytes:             number;
    type:              string;
    etag:              string;
    placeholder:       boolean;
    url:               string;
    secure_url:        string;
    original_filename: string;
}